fix: finalize changes to latest contract code
