+++ src/aibtcdev-dao/extensions/core-proposals/get-voting-power.ts
@@ -1,27 +1,34 @@
 import {
   callReadOnlyFunction,
-  ClarityType,
-  cvToJSON,
+  cvToValue,
   getAddressFromPrivateKey,
+  principalCV,
 } from "@stacks/transactions";
 import { CONFIG, deriveChildAccount, getNetwork } from "../../../utilities";
 
-// gets set protocol treasury in core proposal contract
+// gets voting power for an address on a proposal
+
 async function main() {
   const [
     daoCoreProposalsExtensionContractAddress,
     daoCoreProposalsExtensionContractName,
   ] = process.argv[2]?.split(".") || [];
+  const [daoProposalContractAddress, daoProposalContractName] =
+    process.argv[3]?.split(".") || [];
+  const voterAddress = process.argv[4];
 
   if (
     !daoCoreProposalsExtensionContractAddress ||
-    !daoCoreProposalsExtensionContractName
+    !daoCoreProposalsExtensionContractName ||
+    !daoProposalContractAddress ||
+    !daoProposalContractName ||
+    !voterAddress
   ) {
     console.log(
-      "Usage: bun run get-protocol-treasury.ts <daoCoreProposalsExtensionContract>"
+      "Usage: bun run get-voting-power.ts <daoCoreProposalsExtensionContractAddress> <daoProposalContract> <voterAddress>"
     );
     console.log(
-      "- e.g. bun run get-protocol-treasury.ts ST35K818S3K2GSNEBC3M35GA3W8Q7X72KF4RVM3QA.wed-core-proposals"
+      "- e.g. bun run get-voting-power.ts ST35K818S3K2GSNEBC3M35GA3W8Q7X72KF4RVM3QA.wed-action-proposals ST35K818S3K2GSNEBC3M35GA3W8Q7X72KF4RVM3QA.wed-base-bootstrap-initialization ST35K818S3K2GSNEBC3M35GA3W8Q7X72KF4RVM3QA"
     );
 
     process.exit(1);
@@ -38,20 +45,17 @@ async function main() {
   const result = await callReadOnlyFunction({
     contractAddress: daoCoreProposalsExtensionContractAddress,
     contractName: daoCoreProposalsExtensionContractName,
-    functionName: "get-protocol-treasury",
-    functionArgs: [],
+    functionName: "get-voting-power",
+    functionArgs: [
+      principalCV(voterAddress),
+      principalCV(daoProposalContractAddress),
+    ],
     senderAddress,
     network: networkObj,
   });
 
-  if (result.type === ClarityType.OptionalNone) {
-    console.log("Contract has not been initialized with a treasury contract");
-  }
-  if (result.type === ClarityType.OptionalSome) {
-    const jsonResult = cvToJSON(result);
-    // TODO: might need a better way to display the result
-    console.log(jsonResult);
-  }
+  const parsedResult = cvToValue(result, true);
+  console.log(parsedResult);
 }
 
 main().catch((error) => {
